requirements.txt
numpy
pandas
python-dotenv
python_Levenshtein
scikit_learn
supabase